moved drizzle component to root layout
