--- web/layouts/root.tsx
@@ -1,9 +1,0 @@
-import { Drizzling } from "@/components/drizzling";
-
-export default function Root() {
-  return (
-    <div className="flex min-h-screen items-center justify-center">
-      <Drizzling />
-    </div>
-  );
-}
